arm64: traps: fix compilation error
When enabling verbose printing for kernel OOPS, the stack memory
dump in function dump_backtrace_entry causes compilation failure.
This is because the "stack" variable is not defined.

But, dump_backtrace_entry is meant to print one entry anyway, so
there is no real use for printing memory around that entry. The
memory dump is done in dump_backtrace for each register.

Change-Id: I2ecdb4836b228909bc8c5a896c53fa65d634fd1d
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1657648
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>