hda: enable sor safe clk before sor register read
HDA does a query for the dev_id as a part of the
bottom half interrupt handler for hdmi hotplug.

However if the hotplug handler gets called before
the hdmi enable function turns on the safe clk,
it will cause the sor read to timeout.

Previously the driver unconditionally turned on
clocks during initialization. This is now being
deprecated hence we need to ensure we turn on
all required clocks before writing to a display
resource.

Bug 2022680

Change-Id: I979d50231a8ec194dab764b917855c6cce195582
Signed-off-by: Anshuman Kar <anshumank@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1650826
Reviewed-by: Santosh Galma <galmar@nvidia.com>
Tested-by: Santosh Galma <galmar@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>