bus: tegra-aconnect: fix low code coverage
tegra-aconnect driver is built as part of the kernel image and remove()
does not get called in driver life time and is resulting in low code
coverage. Adding IS_MODULE() checks for remove() apt, so that it gets
used only when the driver is a loadable module.

Bug 200390242

Change-Id: I7f5b3a5d081f6568f8a6cb37d241ccacba3b466c
Reviewed-on: https://git-master.nvidia.com/r/1660793
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>