ALSA: hda/tegra: Fix name lengths
The soundcard driver and short names have been changed for the
Tegra194 Galen platform to be 'tegra-hda-galen-t194' instead of the
default 'tegra-hda'. The driver name has a maximum length of 16
characters and the shortname has a maximum length of 32 characters.
The driver name for Galen is too long and this is causing it to
overflow the driver name buffer and under the procfs the driver
appears as 'tegra-hda-galen-tegra-hda-galen-t194' because there is
no NULL termination and so both the driver name and short name are
read as a single string.

Fix this by ensuring we only copy the first N characters of the name
that can be stored in the driver name and short name buffers.

Bug 200388261
Bug 200382625

Change-Id: I8c221beac092b7a5fc817b8faefe16a3de459ee3
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1661242
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>