t194: hda: property to pass GSC ID value
HDA as a client has read access to APR, which is defined as Carveout-10.
The configuration settings for CO-10 are done in bootloader to allow
access to HDA. Apart from this there is a field in cfg settings for C0-10,
"Allow AID mismatch", this is disabled now. This essentially means the
client requesting for CO-10 access should transmit the same ID for it to
work. This patch helps in passing the ID and same gets programmed in hda
machine driver.

Bug 200371345

Change-Id: I4da3270f69485a82a9c455dde6edb5213f22afab
Reviewed-on: https://git-master.nvidia.com/r/1662248
Reviewed-by: Automatic_Commit_Validation_User
Tested-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>