ALSA: hda: program GSC ID
For HDA as client to have access to APR region, it should transmit the
correct AID(aperture ID or GSC ID). Currently CO-10 is used for APR and
it expects the AID to be matching. So this has to be programmed in
HDA_HDA_GSC_ID_0 (addr: 0x35101e0)

Bug 200371345

Change-Id: I3e162f175afd394da07da1df8726904f7ea034ff
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1662249
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>