media: i2c: Fix IMX274 DOL modes for T19x
This changes the DOL modes to output additional
margin pixels so that we can circumvent VI
alignment limitations for the sake of cropping
LI markers from VI5-deinterleaved DOL captures.

This also fixes the 4k mode vertical resolution
to be 2160 instead of 2166, which was including
margin pixels from the top of the active region
which should have been ignored. The cropping
registers have been adjusted to fix this, too.

Bug 200362372

Change-Id: I4a7a22cf2e993d096834b52158e34c12052817b4
Signed-off-by: Ian Stewart <istewart@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1671540
Reviewed-by: Michelle Bourland <msoult@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Frank Chen <frankc@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>