dc: modify dc resume sequence
The current dc resume sequence for DP and HDMI
schedules a hpd worker in a new thread and then
turns on the DC controller.

This creates a race since tegra_dc_enable should
only be executing after the hpd worker has
completed.

This patch adds a completion variable to the dc
struct which allows for output sources supporting
hotplug to wait for the hotplug worker to complete
during resume.

Bug 20037652
Bug 200354479

Change-Id: Ie23061e0784836534cbc65a228e257f31f237400
Signed-off-by: Anshuman Kar <anshumank@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1666173
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>