dts: t210: apbmisc: Fix ramcode read failure
Memory freq is get stuck in 204Mhz/800Mhz on Darcy(3s memory),
and fail to switch to the other freq.

Applying the incorrect base address(0x7000e864) for strapping options,
which should be 0x70000008 instead. Leading the Darcy to load the
wrong ramcode and loading the untrained emc-table in kernel,
then causing the failure.

The simplest fix, remove this reduncant node and use the node in
soc.dtsi instead, It will be also better for DT maintaining.

Bug 200387701

Change-Id: Ib2281313236076ae5e5672a077d46019e48e39a1
Signed-off-by: Daniel Fu <danifu@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1669279
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>