fs/proc: exclude processes with > 1024 load_avg
The load_avg value can be inflated/deflated for
scheduling purposes. This inflation/deflation
can occur in many ways one of which is non-zero
nice values.To properly assess the workload of
the application exclude inflated (>1024) load_avg
when selecting the "largest_task" task for a
process.

Jira TGFXP-1638

Change-Id: Iaaa70e28e380472965807e6c93d3ee62ae72ed5b
Signed-off-by: Mubushir Rahman <mubushirr@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1669887
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-by: Simo Melenius <smelenius@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bo Yan <byan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>