serial: combined uart: drain write buffer properly
The current start_tx function does not drain the write buffer correctly.
This can cause tasks that are writing to this port to hang when they
wait for the circular buffer to become empty.

Add a check to make sure that the circular buffer will be fully drained
even if it wraps around.

Wake up tasks that are blocked and waiting for the circular buffer to
become empty.

Bug 2079549
Bug 200390528

Change-Id: I019019707640d735036a2f9a8788e8e81888d0e1
Signed-off-by: Mustafa Yigit Bilgen <mbilgen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1671670
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Vignesh Radhakrishnan <vigneshr@nvidia.com>
Reviewed-by: Robert Collins <rcollins@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>