gpu: nvgpu: Abstract PCI header usage
The mclk code for gp106 requires access to the PCI device
ID headers in Linux. This patch makes it so that the mclk
code does not need to directly include the Linux headers.

This allows us to compile and link the dGPU chips code in
the user space unit testing framework.

JIRA NVGPU-525

Change-Id: I89e2fa7fbb3b67f061e026e48a374951e7934aa5
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1673823
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>