hid: jarvis: Remove duplicate sysfs link
Between kernel-3.10 and kernel-4.9, the kernel changed where it
locates input subsystem sysfs nodes within the hid device sysfs
hierarchy. It used to be located outside of the "####:VID:PID:####"
directory and it used to be required to symlink it in, but now it's
located directly within that directory automatically. So the sysfs
link can be removed since it exists already now.

Bug 200389715

Change-Id: I264bbd02f5750d41011afd0819fe29e280cb9117
Signed-off-by: Spencer Sutterlin <ssutterlin@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1680519
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>