drivers: camera: Fix vi5 fops embedded data width
The unit of embed_x register in VI is in bytes,
but the frame width value we got from device
tree is in pixels, therefore, we need to time it
with BPP_MEM (bytes-per-pixel) to get the total
number of bytes for embedded data.

Without this fix, some sensors with longer embedded
data lines will trigger VI errors.

Bug 200389898

Change-Id: I7bbdff77eeb4056cbfa2649688edb59c91237a1a
Signed-off-by: Frank Chen <frankc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1681286
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Vincent Chung <vincentc@nvidia.com>
Reviewed-by: Ruicheng Dai <rdai@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>