drivers: change clock api's of aconnect and adma
The ADMA and Aconnect audio drivers currently uses pm_clk_xxxx
for handling the ahub and ape clocks. In this case the refcount
for these clocks becomes non-zero though the clocks are not enabled.

As per BPMP requirement we need to use devm clock api's.

Bug 200400041

Change-Id: I495cbb242b0713d329f67c2f71d098ac3acb7b86
Signed-off-by: Mohan Kumar <mkumard@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1684236
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>