platform: tegra: add new thermal throttling driver
Add a new throttling driver that can support one or more cooling
devices which can throttle one or more clocks.

Parse the DT to read the clocks that need to be throttled and
calculate the throttling frequency by doing a linear interapolation
between the min and max clock rates.

Add initial support to throttle the CPUs using cpufreq and GPU using
pmqos.

Allow the slope and offset of the throttling curve to be tunable via
DT properties.

Bug 2022953

Change-Id: I744e1e4f65bcf12363e3f9a97cd4aebbf2c6e86a
Signed-off-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1699802
GVS: Gerrit_Virtual_Submit
Reviewed-by: Navneet Kumar <navneetk@nvidia.com>
Reviewed-by: Bo Yan <byan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>