tegra: dc: filter interlaced modes for HDMI and DP
- interlaced modes are supported in hw for T21x and
T18x and T19x but the sw support is not properly
maintained and tested as it was not required for any
of the usecase.
- currently, they are being filtered in graphic
clients like HWC,X etc during mode selection.
- filter out all interlaced modes in kernel itself
to avoid modes that are not properly tested, being
exposed to userspace until sw support is properly
verified.

Bug 200399665

Change-Id: I9901e0adcf03ba2573f66f57fd82cab6a8944efe
Signed-off-by: Santosh Reddy Galma <galmar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1691281
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>