t19x: dts: Add support for audio super-module
Add support for the E2614 super-module for Tegra194. Note that this
will be used by both Galen and Rey and so place this in the common
modules area so that it can be used by both boards. The regulator
nodes for Galen and Rey are different and so the regulator
information will be overlaid on top of this file for each board.

Bug 2101552

Change-Id: If362d64c2ff1b846bb97ce80d0ca5fc1cd7b700c
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1693739
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>