pva: WAR for t194 mailbox DEAD1001 bug
This change works around a HW bug where DEAD1001 cannot be
written to a mailbox. It uses a bidirectional mailbox ( 4 and 6)
as side channel to communicate bit 0 of the actual value of the
mailboxes in use.

Each mailbox is assigned a bit according to its number e.g. mailbox 3
is assigned bit 3.

The use of two side channels is needed to avoid host and FW interfering
with each other.

Channel 4 is used by host write as side channel
Cheanel 6 is used by FW write as side channel

Bug 2090939
Jira PVA-1040

Change-Id: Id57eee9becbb77054ae2464c0b675abd2a5a0150
Signed-off-by: Omar Nemri <onemri@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1700056
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>