ucsi_ccg: USBC PD multiple chargers handling
For CCG4, there are two USB-C ports. The user might connect two
USB-C power adapters simultaneously. If there is already a power
adapter exist on one port, CCG FW would suspend the power path
at the port with newly added power adapter.

But, when the charging power adapter is removed, it's not safe
for CCG FW to enable the power path in the background
(the charger voltage and current rating might be different on
another power adapter).

The goal of this patch is to handle:
1. Parsing the power path information from CCG FW
2. Ignore the power contract completed event when the path is
   being suspended by CCG FW
3. When a power adapter is being removed, check the partner
   on another port. If there is a suspended charger,
   send Get_Source_Cap to trigger the renegotiation for the
   PD contract

Note that it's required to enable the feature with new CCG FW.
But this driver would be compatible with the previous version
of CCG FW.

Also removed ISL OTG control in this patch

bug 200397307
bug 2079219

Change-Id: I68268cd17360e9b932f2db988983688475736f1b
Signed-off-by: Sing-Han Chen <singhanc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1700805
Reviewed-by: WK Tsai <wtsai@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Allie Liu <alliel@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>